# gmmsales-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GMMSales API Documentation",
        "description": "Dokumentasi API untuk aplikasi GMMSales - sistem pencatatan kunjungan customer oleh sales lapangan dengan dokumentasi GPS dan foto.",
        "contact": {
            "name": "Rakha Hafizh",
            "email": "rakhahafizh@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "GMMSales Local Development Server"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Daftar customer milik sales yang login",
                "operationId": "2b23ac696cfd81c23807c2c1fd63b0b0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar customer berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Daftarkan customer baru (Sales)",
                "description": "Sales mendaftarkan customer baru beserta foto kunjungan, koordinat GPS, dan produk yang ditawarkan. Akan otomatis menandai sales sebagai aktif hari ini.",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_customer",
                                    "alamat",
                                    "nomor_telepon",
                                    "latitude",
                                    "longitude",
                                    "product_id",
                                    "photos"
                                ],
                                "properties": {
                                    "nama_customer": {
                                        "type": "string",
                                        "example": "Toko Jaya Abadi"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 10"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.1751
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.865
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "photos[]": {
                                        "description": "1-3 foto (JPG/PNG, max 5MB per file)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer berhasil didaftarkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customers/history": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Riwayat kunjungan customer milik sales yang login",
                "operationId": "074755d656ca2c63dabe47405d13fb9a",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat customer berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Detail customer milik sales yang login",
                "operationId": "d51ce35bfe185b82c242dd4443ffba54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail customer berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Customer bukan milik sales yang login"
                    },
                    "404": {
                        "description": "Customer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Edit data customer sendiri (Sales)",
                "description": "Sales mengubah data tekstual customer miliknya termasuk pilihan produk. GPS dan foto tidak dapat diubah.",
                "operationId": "07a3f33c1373d3c6a97b7d8846d81472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_customer": {
                                        "type": "string",
                                        "example": "Toko Jaya Abadi Updated"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 10A"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "example": "081234567899"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer berhasil diupdate"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan milik sales yang login"
                    },
                    "404": {
                        "description": "Customer tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Hapus customer (soft delete)",
                "operationId": "634844e24d72540f608f7fd297f5d8b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan milik sales yang login"
                    },
                    "404": {
                        "description": "Customer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum tanpa prefix 'Bearer ' (cukup tokennya saja)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}